+++ src/index.ts
@@ -1,8 +1,8 @@
 export * from './components';
-export * from './generated/index.ts';
-export * from './utilities/index.ts';
+export * from './generated/index';
+export * from './utilities/index';
 export * from './factory.tsrx';
-export * from './providers/index.ts';
+export * from './providers/index';
 export {
 	createMachineHook,
 	type MachineHook,